customized webserver and added actions
