+++ .github/actions/build-and-push-container/action.yaml
@@ -1,0 +1,36 @@
+---
+name: Build and Push - Docker
+description: Builds and pushes docker image to Docker Hub via Packer
+inputs:
+  username:
+    required: true
+    description: "Docker Hub username"
+  token:
+    required: true
+    description: "Docker Hub password (PAT)"
+  packer-path:
+    required: true
+    description: "Path to packer file"
+
+runs:
+  using: "composite"
+  steps:
+    - name: Setup `packer`
+      uses: hashicorp/setup-packer@main
+      id: setup
+      with:
+        version: "latest"
+
+    - name: Run `packer init`
+      id: init
+      run: "packer init ${{ inputs.packer-path }}"
+
+    - name: Run `packer validate`
+      id: validate
+      run: "packer validate ${{ inputs.packer-path }}"
+
+    - name: Build Artifact
+      run: "packer build -color=false -on-error=abort ${{ inputs.packer-path }}"
+      env:
+        PKR_VAR_docker_username: ${{ inputs.dockerhub-username }}
+        PKR_VAR_docker_token: ${{ inputs.dockerhub-token }}
