--- .github/actions/build-image.yaml
@@ -1,6 +1,0 @@
----
-# on:
-#   push:
-#     [main]
-
-# build docker image and push to docker hub
