--- .github/actions/build-webserver.yaml
@@ -1,6 +1,0 @@
----
-# on:
-#   push:
-#     [main]
-
-# build webserver and check liveness probe
