+++ .github/actions/scan-container/action.yaml
@@ -1,0 +1,34 @@
+---
+name: Build and Scan Container Image
+
+on:
+  push:
+    branches: [ "main" ]
+  pull_request:
+    branches: [ "main" ]
+
+jobs:
+  build-scan:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+
+      - name: Setup Docker Buildx
+        uses: docker/setup-buildx-action@v2
+
+      - name: Build container
+        run: |
+          docker build ./applications/echo-server --file Dockerfile --tag localbuild/testbuild:latest
+
+      - name: Scan container image
+        uses: anchore/scan-action@v3
+        id: scan
+        with:
+          image: "localbuild/testbuild:latest"
+          fail-build: true
+          severity-cutoff: critical
+      
+      - name: Inspect SARIF report
+        run: cat ${{ steps.scan.outputs.sarif }}
