+++ .github/workflows/build-app.yaml
@@ -1,0 +1,31 @@
+---
+name: Build App
+
+on:
+  push:
+    branches: [ "main" ]
+  pull_request:
+    branches: [ "main" ]
+
+jobs:
+  build-container:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout Repo
+        uses: actions/checkout@v4
+
+      - name: Audit Code
+        uses: ./.github/actions/audit-go-code/action.yaml
+        with:
+          context: "./applications/echo-server/main.go"
+          go-version: ">=1.23"
+
+      - name: Scan container
+        uses: ./.github/actions/scan-container/action.yaml
+
+      - name: Buld and Push to DockerHub
+        uses: ./.github/actions/build-and-push-container/action.yaml
+        with:
+          packer-path: "./applications/echo-server/packer/docker-alpine.pkr.hcl"
+          dockerhub-username: ${{ secrets.DOCKERHUB_USERNAME }}
+          dockerhub-token: ${{ secrets.DOCKERHUB_TOKEN }}
