+++ applications/echo-server/ansible/playbook.yaml
@@ -4,6 +4,11 @@
   become: true
 
   tasks:
+    - name: Install - Go
+      community.general.apk:
+        name: go
+        state: present
+
     - name: Install - perl
       community.general.apk:
         name: perl
