+++ applications/echo-server/html/index.html
@@ -1,0 +1,32 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <title>Go Server</title>
+    </head>
+    <style>
+        html, body {
+            height: 100%
+        }
+
+        html {
+            display: table;
+            margin: auto;
+        }
+        img {
+            display: table;
+            margin: auto;
+            width: 50%;
+        }
+        body {
+            display: table-cell;
+            vertical-align: middle;
+        }
+    </style>
+    <body>
+        <h1>{{.Title}}</h1>
+        <p>{{.Message}}</p>
+        <img src="images/doge.png">
+    </body>
+</html>
