+++ applications/echo-server/packer/variables.pkr.hcl
@@ -1,4 +1,8 @@
-variable "dockerpwd" {
+variable "docker_username" {
+    type = string
+}
+
+variable "docker_token" {
     type      = string
     sensitive = true
 }
